python/requirements.txt

flask==2.0.1
flask-cors==3.0.10
tensorflow>=1.15,<3.0
keras>=2.0.0
numpy>=1.16.0
pillow>=8.0.0
h5py>=2.10.0
